Fixed vafmt/nonvoid test vafmt/nonvoid now works on msvc
--HG--
branch : hookman_autogen
extra : convert_revision : svn%3Ac2935e3e-5518-0410-8daf-afa5dab7d4e3/branches/hookman_autogen%40568

+++ sourcehook/test/testhookmangen.cpp
@@ -469,7 +469,7 @@ namespace
 	THGM_SETUP_PI1(212,
 		int, SourceHook::PassInfo::PassType_Basic, SourceHook::PassInfo::PassFlag_ByVal
 		);
-	THGM_SETUP_RI(212, float, SourceHook::PassInfo::PassType_Float, SourceHook::PassInfo::PassFlag_ByVal);
+	THGM_SETUP_RI(212, double, SourceHook::PassInfo::PassType_Float, SourceHook::PassInfo::PassFlag_ByVal);
 
 	THGM_MAKE_TEST1_vafmt(213, PodRet8, int);
 	THGM_SETUP_PI1(213,
